Add MinGW and MinGW64 support

+++ programs/echo_server_upcall.c
@@ -123,14 +123,14 @@ handle_upcall(struct socket *sock, void *data, int flgs)
 						return;
 					}
 
-				printf("Msg of length %d received from %s:%u on stream %d with SSN %u and TSN %u, PPID %lu, context %u.\n",
+				printf("Msg of length %d received from %s:%u on stream %d with SSN %u and TSN %u, PPID %u, context %u.\n",
 				       (int)n,
 				       namebuf,
 				       port,
 				       rn.recvv_rcvinfo.rcv_sid,
 				       rn.recvv_rcvinfo.rcv_ssn,
 				       rn.recvv_rcvinfo.rcv_tsn,
-				       ntohl(rn.recvv_rcvinfo.rcv_ppid),
+				       (uint32_t)ntohl(rn.recvv_rcvinfo.rcv_ppid),
 				       rn.recvv_rcvinfo.rcv_context);
 				if (flags & MSG_EOR) {
 					struct sctp_sndinfo snd_info;
